Finish basics of Dashboard
