+++ templates/center_dashboard.html
@@ -4,7 +4,6 @@
 {% endblock style %}
 
 {% block main %}
-	<!-- Render flashed messages -->
 <main class="container">
 	<aside class="sidebar">
 		<h2 class="sidebar-heading">{{ center.name }} Information</h2>
@@ -20,6 +19,7 @@
 
 	<div class="content">
 
+		<!-- Render flashed messages -->
 		<div>
 			{% with messages = get_flashed_messages(with_categories=true) %}
 			{% if messages %}
@@ -33,14 +33,17 @@
 		</div>
 
 		<h1 class="stats-title">Blood Bags Statistics :</h1>
-		<a href="{{ url_for('update_bags') }}"><button class="update-button">Update Blood Bags</button></a>
+		<a href="{{ url_for('update_blood_bags') }}"><button class="update-button">Update Blood Bags</button></a>
 
 		<div class="blood-bags">
 			{% for bag in bags %}
 			<article class="blood-bag">
 				<h2 class="blood-category">{{ bag.blood_category }}</h2>
-				<p class="quantity">Quantity: {{ bag.quantity }}</p>
-				<p class="situation">Situation: {{ bag.situation }}</p>
+				<p class="quantity"><strong>Quantity:</strong> {{ bag.quantity }} Bags</p>
+				<p class="situation"><strong>Situation:</strong> {{ bag.situation }}</p>
+				<div class="updated_at">
+					<p>Updated at : {{ bag.updated_at.strftime('%a %d/%m/%Y') }}</p>
+				</div>
 			</article>
 			{% if loop.index % 4 == 0 %}
 		</div><div class="blood-bags">
