+++ templates/layout.html
@@ -48,7 +48,7 @@
 				{% if current_user.city_id %}
 				<a href="{{ url_for('center_dashboard') }}"><button class="header_btn" type="button">Dashboard</button></a>
 				{% else %}
-				<a href="{{ url_for('donor_profile') }}"><button class="header_btn" type="button">Profile</button></a>
+				<a href="{{ url_for('donor_dashboard') }}"><button class="header_btn" type="button">Profile</button></a>
 				{% endif %}
 				<a href="{{ url_for('logout') }}"><button class="header_btn" type="button">Logout</button></a>
 				{% else %}
